Add jump, crouch, and move left/right
int16_t changes needed to prevent negatives turning into ~255
and causing funny polygon bugs. May have gone overboard but
it seems okay; memory changes are nominal.

+++ arduino/any-kind-of-car-and-any-kind-of-truck/utils.h
@@ -4,8 +4,8 @@
 #include "Arduino.h"
 
 struct Xy {
-  uint8_t x;
-  uint8_t y;
+  int16_t x;
+  int16_t y;
 };
 
 void drawLine(
